Version 15
Snap to commit 1265b7e83b43ebc8227d9243591914ea0721cec4

Allow enabling systrace and bugreport captures separately in LoopbackApp.
Capturing a bugreport takes a lot of time and CPU power. The CPU usage
may cause another glitch, or an unrelated glitch may be missed because
the app is waiting for a bugreport to complete.

Capture systrace/bugreport on late callbacks in LoopbackApp.

Remove obsolete method for systrace from native code in LoopbackApp.
The code for this was unused.

Add more atomic methods to audio_utils in LoopbackApp.
Note that these methods are temporary until we can move to the standard
atomic library for C++.

Discard player callbacks if recorder is not yet running.

Display channelIndex in LoopbackApp as MONO when appropriate.

Iteratively calculate variance in LoopbackApp for Java.

Iteratively calculate variance in LoopbackApp for native only.
It is converted into standard deviation and reported by the app.

Refactor statistics recording into a new function in LoopbackApp.
This will make it easier to add new metrics and adjust the existing ones.

Deduplicate code dealing with buffer period stats in LoopbackApp.
Previously, the code for the recorder and player was copied and pasted
with minimal changes.

Fix build warning.

Refactor LoopbackApp buffer stats into one struct.
By unifying the player and recorder stats, copy-pasted code can be
eliminated and adding new statistics is made much easier.

Fix how late buffer callbacks are counted in LoopbackApp.
The previous calculation id not take rounding into account which causes
it to incorrectly label callbacks which occured only 1 ms late as over
the threshold of 2 ms late. This results in the metric being
oversensitive and flaky.

Add setting to LoopbackApp to ignore beginning of recording.
This is useful to work around hardware bugs that cause pops or other
noise at the beginning of a recording.